Flytter logikken for å sjekke om det er dupliserte utbetalinger på barn til selve AutovedtakTjenesten.
