+++ src/main/resources/db/migration/V238__satskjoering_feiltype.sql
@@ -1,0 +1,2 @@
+ALTER TABLE satskjoering
+    ADD COLUMN feiltype VARCHAR DEFAULT NULL;
